Added storage diff
